+++ src/core/Buttons/SplitButton/index.ts
@@ -1,0 +1,3 @@
+export { SplitButton } from './SplitButton';
+export type { SplitButtonProps } from './SplitButton';
+export default './SplitButton';